Move conflict resolution to own file
